Add Ajax unobstrusive

+++ FastService/Controllers/HomeController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
+using Backend;
 
 namespace FastService.Controllers
 {
@@ -35,32 +36,61 @@ namespace FastService.Controllers
             {
                 var user = System.Web.HttpContext.Current.Session["USER"] as string;
 
-                var model = (from x in _dbContext.UsuarioRol
-                             join u in _dbContext.Usuario on x.UserId equals u.UserId
-                             join r in _dbContext.Role on x.RolId equals r.RolId
-                             where u.Email == user
-                             select new MenuModel()
-                             {
-                                 Nombre = u.Nombre,
-                                 Apellido = u.Apellido,
-                                 DefaultController = r.DefaultController,
-                                 DefaultAction = r.DefaultAction,
-                                 MenuItems = (from rm in _dbContext.RoleMenu 
-                                              join i in _dbContext.ItemMenu on rm.ItemMenuId equals i.ItemMenuId
-                                              where rm.RolId == x.RolId
-                                              select new MenuItemModel()
-                                              {
-                                                  MenuId = i.ItemMenuId,
-                                                  MenuName = i.Name,
-                                                  DisplayName = i.Name,
-                                                  Controller = i.Controlador,
-                                                  Action = i.Accion,
-                                                  ParentId = i.ItemMenuPadreId
-                                              }).ToList()
-                             }).ToList().FirstOrDefault();
 
-                result = this.View(model);
+                if (CommonUtility.IsDevelopmentServer())
+                {
+                    var model = new MenuModel()
+                    {
+                        Nombre = "Leandro",
+                        Apellido = "Lopez",
+                        DefaultAction = "Create",
+                        DefaultController = "Payment",
+                        MenuItems = new List<MenuItemModel>()
+                        { new MenuItemModel()
+                            {
+                                Action = null, Controller = null,DisplayName = "Mercado Pago",MenuId = 1,MenuName = "MercadoPago",ParentId = null
+                            },
+                            new MenuItemModel()
+                            {
+                                Action = "Create", Controller = "Payment",DisplayName = "Nuevo Pago",MenuId = 1,MenuName = "NuevoPago",ParentId = 1
+                            }
+                    }};
+
+                    result = this.View(model);
+                }
+                else
+                {
+
+                    var model = (from x in _dbContext.UsuarioRol
+                                 join u in _dbContext.Usuario on x.UserId equals u.UserId
+                                 join r in _dbContext.Role on x.RolId equals r.RolId
+                                 where u.Email == user
+                                 select new MenuModel()
+                                 {
+                                     Nombre = u.Nombre,
+                                     Apellido = u.Apellido,
+                                     DefaultController = r.DefaultController,
+                                     DefaultAction = r.DefaultAction,
+                                     MenuItems = (from rm in _dbContext.RoleMenu
+                                                  join i in _dbContext.ItemMenu on rm.ItemMenuId equals i.ItemMenuId
+                                                  where rm.RolId == x.RolId
+                                                  select new MenuItemModel()
+                                                  {
+                                                      MenuId = i.ItemMenuId,
+                                                      MenuName = i.Name,
+                                                      DisplayName = i.Name,
+                                                      Controller = i.Controlador,
+                                                      Action = i.Accion,
+                                                      ParentId = i.ItemMenuPadreId
+                                                  }).ToList()
+                                 }).ToList().FirstOrDefault();
+
+                    result = this.View(model);
+
+                }
+
             }
+
             return result;
         }
